Moving things around and Polysemy
