+++ lib/Entry.hs
@@ -1,12 +1,12 @@
 {-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE TemplateHaskell #-}
 
-module Vault.Entry
+module Entry
   ( newFieldKey
   , dateModified
   , Entry (..), EntryConvertible (..), emptyEntry
-  , path, fields
-  , Field, emptyField
+  , path, masterField, fields
+  , Field (..), emptyField, getFieldKey
   , value
   )
 where
@@ -28,9 +28,9 @@ newtype FieldKey = FieldKey T.Text
 instance Hashable FieldKey
 
 instance A.ToJSON FieldKey where
-  toJSON (FieldKey text) = A.toJSON text
-
 instance A.ToJSONKey FieldKey where
+instance A.FromJSON FieldKey where
+instance A.FromJSONKey FieldKey where
 
 newFieldKey :: T.Text -> Maybe FieldKey
 newFieldKey t =
@@ -40,11 +40,15 @@ newFieldKey t =
     Nothing
   where allowedChars = ['a'..'z'] ++ ['A'..'Z'] ++ ['0'..'9'] ++ "-_;:"
 
+getFieldKey :: FieldKey -> T.Text
+getFieldKey (FieldKey t) = t
+
 data Field =
   Field
   { _fDateModified :: DateTime
   , _value :: T.Text
   }
+  deriving (Show, Eq)
 
 -- TODO me no likey, better way? https://github.com/ekmett/lens/issues/286
 emptyField :: Field
@@ -60,8 +64,10 @@ data Entry =
   Entry
   { _path :: [T.Text]
   , _eDateModified :: DateTime
+  , _masterField :: (FieldKey, Field)
   , _fields :: HS.HashMap FieldKey Field
   }
+  deriving (Show, Eq)
 
 -- TODO me no likey, better way? https://github.com/ekmett/lens/issues/286
 emptyEntry :: Entry
@@ -69,10 +75,11 @@ emptyEntry =
   Entry
   { _path = []
   , _eDateModified = ""
+  , _masterField = undefined
   , _fields = HS.empty
   }
 
-makeLensesFor [("_path", "path"), ("_fields", "fields")] ''Entry
+makeLensesFor [("_path", "path"), ("_masterField", "masterField"), ("_fields", "fields")] ''Entry
 
 class DateModified a where
   dateModified :: Lens' a DateTime
@@ -84,4 +91,4 @@ instance DateModified Field where
   dateModified = lens _fDateModified (\e d -> e { _fDateModified = d } )
 
 class EntryConvertible a where
-  entry :: Lens Entry (Maybe Entry) a a
+  entry :: Prism a a Entry Entry
